Refactored files and packages after removing others
